pyproject.toml
[project]
name = "stactools-uvx"
version = "0.1.0"
description = "An application for generating STAC metadata with any stactools package"
readme = "README.md"
authors = [
    { name = "hrodmn", email = "henry@developmentseed.org" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.0",
    "stac-pydantic>=3.2.0",
]

[project.scripts]
stactools-uvx = "stactools_uvx:app"

[project.optional-dependencies]
lambda = [
    "mangum>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
deploy = [
    "aws-cdk-lib>=2.186.0",
    "pydantic-settings[yaml]>=2.8.1",
]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "uvicorn>=0.34.0",
]

[tool.pytest.ini_options]
addopts = "-vv --ignore=cdk.out"

[tool.ruff]

[tool.ruff.lint]

